+++ packages/devextreme/js/__internal/ui/diagram/ui.diagram.context_toolbox.ts
@@ -7,7 +7,8 @@ import type { ShownEvent } from '@js/ui/popover';
 import type { WidgetProperties } from '@ts/core/widget/widget';
 import Widget from '@ts/core/widget/widget';
 import { getDiagram } from '@ts/ui/diagram/diagram.importer';
-import Popover from '@ts/ui/popover/m_popover';
+
+import Popover from '../popover/popover';
 
 const DIAGRAM_CONTEXT_TOOLBOX_TARGET_CLASS = 'dx-diagram-context-toolbox-target';
 const DIAGRAM_CONTEXT_TOOLBOX_CLASS = 'dx-diagram-context-toolbox';
